+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Version 0.11
 ------------
+- function ddlx_get_dependants_recursive() rolled into ddlx_get_dependants()
 
 Version 0.10
 ------------